pyproject.toml
[tool.poetry]
name = "swarmpy"
version = "0.0.0"
description = "A Python library for the ftSwarm Python API"
authors = ["bergschneider <magiclake@gmx.de>"]
readme = "README.md"
license = "MIT"
packages = [
    { include = "swarm" },
]

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
